Improve feature setter robustness
This adds a test to check that all feature.set_value() calls will cause a query, i.e.,
that there are no self.call()s that are not awaited, and fixes existing code in this context.

This also fixes an issue where it was not possible to print out the feature if the value threw an exception.