Handle network key type
Aims to handle network key type by using a text match that assigns the integer, as reported in https://github.com/python-kasa/python-kasa/issues/1325

One solution of multiple potential solutions. Adds the default of 3, which should be WPA2, and is set as a default in multiple other locations within the code.

May need more documentation on the levels (0-4) and their correspondence.