Add new subcommand to monitor plug power
Add a new subcommand to monitor power from kasa power strips in
realtime polling the strip ~1/sec and printing the power useage
of the device and each plug to the terminal for monitoring.

This subcommand mimics the design of several existing Linux based
tools for monitoring system stats and extends it to monitoring
device power.